sync_media.sh: add a "-p" option (for "pretend")

+++ sync_media.sh
@@ -5,6 +5,17 @@ DIRS="/home/marcec/Media
 /home/marcec/Other"
 TARGET="/media/MARCEC_BACKUP/Informal/"
 RSYNC_OPTIONS="-aX --exclude=lost+found/ --delete --delete-excluded --numeric-ids"
+PRETEND=0
+
+while getopts p a
+do
+    case $a in
+        p) PRETEND=1;;
+        *) echo "Invalid option \"$a\"" >&2;
+            exit;;
+    esac
+done
+shift $(expr $OPTIND - 1)
 
 if [ ! -d "$TARGET" ];
 then
@@ -14,6 +25,8 @@ fi
 
 echo "Target is \"$TARGET\"."
 
+[ $PRETEND -eq 1 ] && echo "Would execute:"
+
 echo "$DIRS" | while read d;
 do
     if [ ! -d "$d" ];
@@ -22,6 +35,10 @@ do
         exit
     fi
 
-    # echo rsync "$RSYNC_OPTIONS" "$d" "$TARGET"
-    rsync $RSYNC_OPTIONS "$d" "$TARGET"
+    if [ $PRETEND -eq 1 ]
+    then
+        echo rsync "$RSYNC_OPTIONS" "$d" "$TARGET"
+    else
+        rsync $RSYNC_OPTIONS "$d" "$TARGET"
+    fi
 done
